+++ index.js
@@ -335,6 +335,8 @@ module.exports = function(options) {
       throw "Specify the `shortName` option and set it to the name of your project's repository or folder";
     }
     self.title = self.options.title;
+    // For devops purposes
+    self.options.baseUrl = process.env.APOS_BASE_URL || self.options.baseUrl;
     self.baseUrl = self.options.baseUrl;
     self.prefix = self.options.prefix || '';
   }